+++ stasis-ios/Stasis/StasisError.m
@@ -25,7 +25,7 @@
 
 - (NSString *)localizedDescription
 {
-	NSString *key = [NSString stringWithFormat:@"%@#%d", self.domain, self.code];
+	NSString *key = [NSString stringWithFormat:@"%@#%ld", self.domain, (long)self.code];
 	NSBundle *bundle = [NSBundle bundleWithURL:[[NSBundle mainBundle] URLForResource:@"StasisResources" withExtension:@"bundle"]];
 	NSString *localizedDescription = [bundle localizedStringForKey:key value:@"" table:@"StasisErrors"];
